Move the old disaggregate submodule to a legacy submodule.
Modules, tests, notebooks and README.md were updated.
