+++ README.md
@@ -9,8 +9,12 @@ itemised energy bill from just a single, whole-house power meter.
 
 NILMTK is a toolkit designed to help **researchers** evaluate the accuracy of NILM algorithms. 
 
-**As of June 2018, NILMTK is being revived! Although no major changes are expected in the coming months, the codebase is slowly being updated to work properly with the current Python ecosystem, especially to modern versions of our major dependency, *Pandas*. 
-It may take time for the NILMTK authors to get back to you regarding queries/issues. However, you are more than welcome to propose changes, support!** 
+**⚠️It may take time for the NILMTK authors to get back to you regarding queries/issues. However, you are more than welcome to propose changes, support!** 
+
+**⚠️The last steps in the development of the next NILMTK version (0.4) are being done in the master git branch. If you are a new user, please refer to the version 0.3.x branch at [https://github.com/nilmtk/nilmtk/tree/0.3.x](https://github.com/nilmtk/nilmtk/tree/0.3.x).**
+
+**⚠The latest release is available through [https://github.com/nilmtk/nilmtk/releases/latest](https://github.com/nilmtk/nilmtk/releases/latest)**
+
 
 # Documentation
