+++ nilmtk/legacy/disaggregate/combinatorial_optimisation.py
@@ -5,10 +5,10 @@ import numpy as np
 import pickle
 import copy
 
-from nilmtk.utils import find_nearest
-from nilmtk.feature_detectors import cluster
-from nilmtk.disaggregate import Disaggregator
-from nilmtk.datastore import HDFDataStore
+from ...utils import find_nearest
+from ...feature_detectors import cluster
+from . import Disaggregator
+from ...datastore import HDFDataStore
 
 
 class CombinatorialOptimisation(Disaggregator):
