+++ nilmtk/legacy/disaggregate/disaggregator.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from nilmtk.timeframe import merge_timeframes, TimeFrame
+from ...timeframe import merge_timeframes, TimeFrame
 
 
 class Disaggregator(object):
