+++ nilmtk/tests/test_metrics.py
@@ -6,7 +6,7 @@ from nilmtk import (Appliance, MeterGroup, ElecMeter, HDFDataStore,
 from nilmtk.utils import tree_root, nodes_adjacent_to_root
 from nilmtk.elecmeter import ElecMeterID
 from nilmtk.building import BuildingID
-from nilmtk.disaggregate import CombinatorialOptimisation
+from nilmtk.legacy.disaggregate import CombinatorialOptimisation
 from nilmtk.metrics import f1_score
 
 class TestMetrics(unittest.TestCase):
